+++ src/odemis/gui/cont/acquisition/__init__.py
@@ -1,0 +1,39 @@
+# -*- coding: utf-8 -*-
+"""
+Created on 22 Aug 2012
+
+@author: Éric Piel, Rinze de Laat, Philip Winkler
+
+Copyright © 2012-2022 Éric Piel, Rinze de Laat, Delmic
+
+This file is part of Odemis.
+
+Odemis is free software: you can redistribute it and/or modify it under the
+terms of the GNU General Public License version 2 as published by the Free
+Software Foundation.
+
+Odemis is distributed in the hope that it will be useful, but WITHOUT ANY
+WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
+PARTICULAR PURPOSE. See the GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License along with
+Odemis. If not, see http://www.gnu.org/licenses/.
+
+
+### Purpose ###
+
+This module contains classes to control the actions related to the acquisition
+of microscope images.
+
+"""
+
+from .auto_center import AutoCenterController
+from .cryo_acq import CryoAcquiController
+from .cryo_z_localization import CryoZLocalizationController
+from .fastem_acq import (FastEMAcquiController, FastEMCalibrationController,
+                         FastEMOverviewAcquiController, FastEMScintillatorCalibrationController)
+from .fine_align import FineAlignController
+from .overview_stream_acq import OverviewStreamAcquiController
+from .secom_acq import SecomAcquiController
+from .snap_shot import SnapshotController
+from .sparc_acq import SparcAcquiController
